[remix] Enable `nativeFetch` when `v3_singleFetch` future flag is enabled (#12918)
`nativeFetch` already is enabled when the previously named
`unstable_singleFetch` future flag is enabled, so do the same for the
stabalized flag name `v3_singleFetch`.

For more info, see the **"Deprecated `fetch` polyfill"** section of the
[Remix
docs](https://remix.run/docs/en/main/guides/single-fetch#enabling-single-fetch):

> Single Fetch requires using `undici` as your `fetch` polyfill, or
using the built-in `fetch` on Node 20+, because it relies on APIs
available there that are not in the `@remix-run/web-fetch` polyfill …
you will need to call `installGlobals({ nativeFetch: true })` to use
undici.