[remix] extract common edge file tracing logic (#12953)
This PR extracts common logic between the node file trace logic for edge
functions in the React Router and Remix createRender functions.

This PR is minimal refactor that extracts the logic, in a future PR it
may be worth creating an "Edge File Trace" function that completely
encapsulates this logic.