[next][tests][examples] pin fixtures and set up auto updates (#12979)
Having these tests run against `next@canary` and `react@latest` allows
changes in those projects to break the tests unpredictably. This happens
somewhat often, disrupting other work in this repo that happens to touch
the builder interface or some other global invalidation of the turbo
cache.

In order to bring those breaks to a predictable (and singular) place,
this PR pins the Next versions to their current resolved values. There
is already a cron set up to update these, which has been fixed in this
PR to work more completely.

After this merges, when a new version of `next@canary` is published, a
new PR will be opened to this repo with that change, scoping any breaks
to that PR.

---

**Key Changes:**

- split out the cron updated into two parts
- .github/workflows/cron-update-next-canary.yml: update all of the
`packages/next` canary fixtures
- .github/workflows/cron-update-next-latest.yml: update the uses in
other packages to latest
- utils/update-next.js
- fixed a bug that prevented deep traversal when looking for fixtures to
update
    - generalized to allow for canary or latest updates
- ran this updated script to produce most of the fixture changes in this
PR
- removed `ignoreNextjsUpdates` from most next fixtures, but left it
where it appeared we wanted to test specific old versions

---

**Considerations:**

- should this cron also pin `react@latest` to its resolved value?
- see the author's comments in this PR

---------

Co-authored-by: Austin Merrick <onsclom@onsclom.net>