[remix] Leverage project settings to determine framework (#13056)
Turns out that the `react-router.config.*` file is optional, so it's not
a complete test to determine if React Router is being used.

The Framework detection logic already considers if the React Router Vite
plugin is being used in `vite.config.*`, so we'll have the Remix builder
rely on the Framework preset being set, which centralizes the detection
logic / removes the (incomplete) duplicated logic.