[errors] improve 'not found' error messaging to include scope (#13111)
There's some spots, currently, that obfuscate that you need to provide
the `--scope` to the command, namely `vc remove` and `vc alias`. We _do_
have good messaging, but it wasn't being caught because of incorrect
checks. Now, when the user doesn't provide the scope when they need to,
we tell them that the deployment under scope `scope name` can't be
found, rather than just the deployment