[cli] get maxDuration from function configuration (#13141)
Before this PR, when we run a `waitUntil` code it's fixed to 10ms max
execution time.

After this PR, the max duration associated to `waitUntil` is read from
teh function configuration, that can be declared at vercel.json, for
example:

```json
{
  "redirects": [
    {
      "source": "/",
      "destination": "https://github.com/nicoalbanese/ai-sdk-slackbot"
    }
  ],
  "functions": {
    "api/events.ts": {
      "maxDuration": 60
    }
  },
  "outputDirectory": "dist"
}
```

So this PR is replacing the fixed value for reading the value from the
configuration, and in case the value si not there, it's using the 30s as
value which is the up to date value.