Add authorization step to cli integration resource creation (#13192)
This work follows up #13154 to allow resource creation to use the CLI
wizard again. All resource creation for integrations were forced to the
Web UI due to all requests failing without the authorization step. Now,
with this PR, we follow the steps for automated authorization, and send
the user to an authorization page on failure (to change the payment
method) or when action is required (such as 3DS).

This PR requires an update to the frontend to enable redirect to the new
manual authorize steps page.

Additionally, while fixing this bug, I found a separate bug revealed by
trying to provision a Supabase resource. It turns out our system wasn't
respecting default values of hidden metadata fields, so Supabase was
returning an error due to missing metadata content.