[next] Enable blocking fallback routes to provide fallback root params (#13215)
This makes it so that blocking fallback routes when PPR is enabled
(those routes with unknown root params) can provide the right metadata
to the prerender configuration. Previously the old code created route
shells for each page when the fallback root shell was used, this now
just generates new root shells to be used.