[cli] refresh `VERCEL_OIDC_TOKEN` (#13226)
If `vc dev` is left running for longer than 12 hours, the
`VERCEL_OIDC_TOKEN` environment variable can expire, leading to errors
in applications that depend on it.

This PR updates `vc dev` to proactively fetch a new `VERCEL_OIDC_TOKEN`
15 minutes before its expiry and restart the dev server, bypassing this
issue. This is a simpler approach than the one in
https://github.com/vercel/vercel/pull/13214, which updated `.env.local`
on disk.

Note that if OIDC is disabled when `vc dev` is run or if OIDC is
disabled at the time of a refresh, then refreshes will be disabled.
Additionally, care is taken not to refresh `VERCEL_OIDC_TOKEN` too
frequently (at most once every minute). This should not be a problem in
practice, because the tokens issued in development have a 12 hour
expiration.

For manual testing, I set the environment variables

```sh
export REFRESH_VERCEL_OIDC_TOKEN_BEFORE_EXPIRY_MILLIS=12h
export REFRESH_VERCEL_OIDC_TOKEN_THROTTLE_MILLIS=10s
```

to force `vc dev` to consider whatever token was fetched already expired
and to speed up fetching the next one.