[next] Support blocking static shells (#13272)
Support blocking static shells properly for PPR. A blocking page is when
there is no fallback configured in the manifest and therefore shouldn't
provide a fallback in those cases. This also modifies the allow query
logic so that it can handle this case even when fallback root params are
provided.