[next] Fix: Segment prefetching when PPR disabled (#13275)
When PPR is disabled for a route, and the route is not part of the
prerender manifest, we must reroute it to an RSC request for the entire
page. Next.js will output this information into the routes manfiest, and
already handles this correctly for dynamically-routed routes, but not
for statically-routed routes.

This updates the builder to handle statically-routed routes, too, using
the same logic that already exists for dynamically-routed routes.

See https://github.com/vercel/next.js/pull/78568 for more context.