feat(auth): implement token refresh (#13317)
This PR implements an automatic token refresh mechanism, a follow-up on
#12098.

During initial login, we will save new tokens and their expiry in the
auth config file.

Then, during API requests, we will insert the following logic:

```mermaid
graph TD;
    A[Has access_token expired?]-->|Yes| B;
    A-->|No| C[Continue];
    B[Has refresh_token expired?]-->|Yes| D;
    B-->|No| E;
    D[Delete auth config];
    D--> C;
    E[Perform token refresh]-->|Successful| F;
    E-->|Unsuccessful| D
    F[Update auth config]-->C
```

This will ensure that if a request can be made, we proceed as usual,
otherwise, the user gets logged out, so as not to leave them in an
undesirable state.

I have refactored the token verification code so we can easily swap it
out with a [OAuth Token
Introspection](https://datatracker.ietf.org/doc/html/rfc7009) endpoint
later on, making the use of JWT purely an implementation detail.