[cli] Warn if system env vars are not available (#13357)
Features like skew protection are not supported without system
environment variables, so `vc build` should warn when running outside of
Vercel. [Linear
issue](https://linear.app/vercel/issue/FLOW-3548/could-we-add-a-warning-when-vc-build-is-run-outside-of-vercel-to).

The `VERCEL=1` env var gets set by the cli in all contexts, so we can't
use that. I used `VERCEL_BUILD_IMAGE` because it is defined in the build
container.